ExecutionFlowRecorder: Fixed mini GC Storm
I know this is going to be replaced by the CFG CPU, but:

We cache the action in a static field, this allows the onReached action
to be a static lambda. This avoid a lot of Actions being created on the
GC Heap to be reclamed by the GC later.

We also don't pass the AddressBreakPoint as a parameter. We discard the
paramater ("_") character, since we are creating the breakpoint.

Signed-off-by: Maximilien Noal <noal.maximilien@gmail.com>