Fix program size
The size of the program image might not match the full size of the exe
file.

Calculate the correct size using the information in the MZ header.

Use the formula at
https://moddingwiki.shikadi.net/wiki/EXE_Format#Machine_code ,
except that (as opposed to what said in that page) the result is the
total size of header + code. To get the pure code length we still must
subtract the header size.