GdbCommandHandler: reply to extended mode request with error message (#1072)
If gdb tries to connect to Spice86 using `extended-remote` (instead of just
`remote`), it will send a `!` command (request to activate extended
mode).

The default answer of Spice86 is an empty response, which is not
acknowledged by the gdb client.

Change the command handler so that it outputs an error packet (with an
arbitrary error code). In this way, gdb recognizes that the extended
mode is not supported, and does not try further.

This fixes the connection problems with Seer, which uses extended-remote
to connect.