Register the int21h "set time" handler in order
I was looking at the int21h services to try to figure out if I needed to
implement a new one that isn't currently supported or not. It turns out
that I don't. But I noticed that they were all in order except for one:
0x2D "set time". It through me off for a second, so I rearranged it.

This commit doesn't change any functionality. It just satisfies my sense
of tidiness.