Add further commentary on how we setup the first MCB in DosMemoryManager
I spent a while looking at DosMemoryManager and trying to figure out
what it was doing while I was tracking down a memory corruption bug. I
thought that there was something wrong with DosMemoryManager itself, but
it turns out that there wasn't. That said, as someone who wasn't
familiar with DOS memory management or memory layout, the minimal
comments that were already there pointed me in the right direction, but
they weren't sufficient to explain why it's being done the way that it
is. I had to dig deeper for that. This commit adds the comments I wish
I'd had at the beginning.