Add unit tests for DosMemoryManager and fix a small off-by-one bug
In the process of trying to track down a suspected memory corruption
issue, I decided to add full unit tests for the DosMemoryManager to
help. I refactored it slightly by adding a new interface to decouple it
from DosProcessManager so that I didn't have to construct as many mocks
to construct a DosProcessManager instance to unit test it.

The unit tests showed that DosMemoryManager pretty much worked correctly
already. There was a small off-by-one bug when modifying a memory block
that I found and fixed, but it's an edge case that it's unlikely a real
game would hit. Apart from that, there are no other functional changes
in this commit.

Unfortunately I didn't find the bug I was looking for, but I think that
these unit tests are valuable anyway. They verify that DosMemoryManager
behaves correctly. Hopefully they'll help us in the future.