Fix memory corruption after DOS int21h 0x48 memory allocation
I have been slowly but surely trying to track down a problem that caused
a game I've been trying to run in Spice86 to crash shortly after it read
the contents of a file into a block of memory that it asked DOS to
allocate for it. Since that file is loaded by the game very early, it
crashed almost immediately. Using the built-in debugger in Spice86, I
managed to track it down to the fact that some of the game code itself
was being corrupted, including the function that loads the file since it
is the first one defined in the executable. I learned a lot about how
DOS process management and memory management works, and how Spice86 does
things. That took me a while.

Eventually I figured out that the core problem is that Spice86 loads the
executable into memory and sets up a stack for it, but it doesn't tell
the DosMemoryManager about it, so DosMemoryManager thinks that, that
memory is still available to allocate, and it happily hands it out the
first time someone asks. The game I was trying to run opens a file, gets
the size of it, allocates a buffer for it, and reads the file into that
buffer. With this bug, the buffer overlaps with where Spice86 loaded the
executable itself, so reading the file into it overwrites the very code
that's reading it, which means that the int21h 0x3F interrupt handler
that reads the file into the buffer returns to a location that no longer
contains valid instructions for the CPU to execute, and Spice86 crashes
shortly thereafter when it dutifully tries to execute them anyway.

I think that the feature proposed in PR #871 "feature: DOS LOAD AND/OR
EXEC and DOS process management" is a step in the right direction
towards solving this properly. However, I tried the code in that PR, and
it does not actually solve this bug as-is. It just introduces a better
structure for solving it by making DOS handle process loading rather
than bypassing it like we do now, so we'd theoretically have the
DosMemoryManager available when we load the process to inform it.

Without that PR, I had trouble figuring out where to put this fix. The
obvious place to put it would be in DosProcessManager.SetupCpuForExe()
when we load the executable (and to do something similar in
DosProcessManager.LoadComFile()), but unfortunately DosProcessManager is
a dependency for DosMemoryManager, so it must be constructed first in
the Dos class's constructor. I didn't want to introduce a circular
dependency or heavily refactor the code like PR #871 is doing, so I
eventually decided to create a new function in ProgramExecutor that
tells DosMemoryManager to reserve the space after the executable is
loaded since it already has knowledge of all of Dos, and it already uses
it to orchestrate loading the executable into memory. However during the
initial code review, Kevin Ferrare suggested putting it in
DosMemoryManager since it so heavily relates to it, and just using
ProgramExecutor to orchestrate it. That allowed me to more easily unit
test it, so that's where I ultimately decided to put it.

My fix is a little ugly, but I did my best to cleanly separate it out
into its own function, following the existing conventions, and document
it well.

I don't have unit tests that actually test loading a full EXE since that
would involve setting up a hefty portion of Spice86 just to load it and
verify that the space was reserved correctly, but I did add unit tests
for the vast majority of the logic - all of the real work that's done in
DosMemoryManager. It's just the tiny bit of orchestration that calls
into the new function in DosMemoryManager in ProgramExecutor that isn't
tested. That seemed like a fair trade. Additionally, all of the existing
unit tests still pass with this fix.

Additionally, I tested this fix manually with my game. It still isn't
playable yet, but it gets *much* further, and Spice86 no longer crashes
when I run it.