Refactor DosProcessManager by splitting out PSP tracking into its own class
This is prep work for supporting proper DOS program load/exec. It
doesn't actually add any new features. I just pulled all of the
properties and functions that deal directly with loading the initial
program start segment and PSP addresses from the configuration and
keeping track of the current PSP segment address from the
DosProcessManager class into the new DosProgramSegmentPrefixTracker
class. Now DosMemoryManager uses it instead of the places where it
previously used DosProcessManager, and DosProcessManager uses it in
place of its prior internal PSP-related methods. This will allow us to
pass DosMemoryManager to DosProcessManager in the future without
creating a dependency cycle.

As a bonus, although this refactor doesn't add any new features or fix
any bugs, it does add unit tests for the new
DosProgramSegmentPrefixTracker. Since DosProcessManager isn't unit
tested, this improves our test coverage for that section of the
codebase. It's not much, but it's a nice side effect.