Add unit tests for DosMemoryManager and fix a small off-by-one bug
In the process of trying to track down a suspected memory corruption
issue, I decided to add full unit tests for the DosMemoryManager to
help.

The unit tests showed that DosMemoryManager pretty much worked correctly
already. There was a small off-by-one bug when modifying a memory block
that I found and fixed. It is only triggered when a program tries to
allocate the entire conventional memory space using its own math rather
than relying on the memory manager to tell it how much is free. Since
some DOS games do this, there was a workaround in the previous version
of DosMemoryManager that tried to compensate for it by subtracting 1
paragraph from the requested block size, but that could lead to
over-allocation and memory corruption. I removed that workaround and
fixed the problem at its source by accounting for the size of the final
paragraph in the memory available for allocation in the DosMemoryManager
constructor. There are now unit tests specifically for this edge case.

I also discovered that if DosMemoryManager was asked to modify a block
to be bigger than the available memory, it would resize it to use all of
the remaining memory, then leave it at that size when it returned. That
would cause problems with programs that used that technique to find the
size of the largest free block. There would never be one, because the
last block would take all of the remaining memory. I fixed that bug so
that TryModifyBlock() always leaves the block at its original size if it
fails. I also added a unit test case to prove that it behaves correctly
in this case.

Unfortunately I didn't find the bug I was looking for, but I think that
these unit tests are valuable anyway. They verify that DosMemoryManager
behaves correctly. Hopefully they'll help us in the future.