Allocate memory for EXE files that are loaded into memory
I have been slowly but surely trying to track down a problem that caused
a game I've been trying to run in Spice86 to crash shortly after it read
the contents of a file into a block of memory that it asked DOS to
allocate for it. Since that file is loaded by the game very early, it
crashed almost immediately. Using the built-in debugger in Spice86, I
managed to track it down to the fact that some of the game code itself
was being corrupted, including the function that loads the file since it
is the first one defined in the executable. I learned a lot about how
DOS process management and memory management works, and how Spice86 does
things. That took me a while.

Eventually I figured out that the core problem is that Spice86 loads the
executable into memory and sets up a stack for it, but it doesn't tell
the DosMemoryManager about it, so DosMemoryManager thinks that, that
memory is still available to allocate, and it happily hands it out the
first time someone asks. The game I was trying to run opens a file, gets
the size of it, allocates a buffer for it, and reads the file into that
buffer. With this bug, the buffer overlaps with where Spice86 loaded the
executable itself, so reading the file into it overwrites the very code
that's reading it, which means that the int21h 0x3F interrupt handler
that reads the file into the buffer returns to a location that no longer
contains valid instructions for the CPU to execute, and Spice86 crashes
shortly thereafter when it dutifully tries to execute them anyway.

I think that the feature proposed in PR #871 "feature: DOS LOAD AND/OR
EXEC and DOS process management" is a step in the right direction
towards solving this properly. However, I tried the code in that PR, and
it does not actually solve this bug as-is. It just introduces a better
structure for solving it by making DOS handle process loading rather
than bypassing it like we do now, so we'd theoretically have the
DosMemoryManager available when we load the process to inform it.

To fix it, I added support to DosMemoryManager for reading the EXE
header and allocating the requested amount of space. Right now this is
done after creating the PSP, but before loading the program image. It is
also only handled for EXE files. It should be done for COM files as
well, but that will need to happen at a later date. This is a step in
the right direction that will make proper DOS LOAD/EXEC support much
easier to implement, and it allocates memory properly for EXE files that
are loaded by Spice86 in the meantime.

By putting the bulk of the logic in DosMemoryManager, I was able to able
to add unit tests for it to the existing DosMemoryManager unit test
suite. I used those tests to rigorously ensure that DosProcessManager
correctly reads the EXE header and handles all documented scenarios that
would be handled by DOS. It's more complicated that I originally thought
it would be.

Additionally, I tested this fix manually with my game. It still isn't
playable yet, but it gets *much* further, and Spice86 no longer crashes
when I run it.