Show block address in hex in DOS INT21H memory MODIFY/FREE
While debugging some memory-related issues, I noticed that the
DosInt21Handler printed out the block segment given for the MODIFY and
FREE operations in decimal rather than in hex. There isn't technically a
problem with that. It works, and it gives the right information. However
almost everywhere else in the code segments are printed in hex. This
commit makes the DOS INT21H memory MODIFY/FREE log messages consistent
with how those addresses are printed elsewhere in Spice86.