fix: add fuzzy file searching
American Gladiators by GameTek request to open a file named "GLAD    .000", which won't be found by Spice88.

DosBox resolves this issue by padding both, the local and requested filename to match the 8.3 specification and then compares both.

This resolves the issue. For consistency, I implemented it to the other search functions.