Add support for saving the video card info to a file
The builtin debugger is able to display a lot of information about the
video card state, which is very useful for tracking down video-related
issues. However prior to this commit, wasn't possible to easily copy it
and compare snapshots of the video card state at different points in
time without manually copying each field or transcribing it in some
other way, which was painful and time-consuming. Therefore this commit
adds a new context menu to the video debugger view that allows the user
to save the whole video card state to disk as a JSON file. That makes it
possible to run through a JSON formatting tool like jq and diff it much
more easily.