Respect the VGA Vertical Sync End register's Write Protect bit
According to the VGA spec, the Vertical Sync End register in the CRT
Controller is supposed to prevent CRT Controller registers 0 - 7 from
being written. This bit was unused in the earlier EGA spec. This bit is
set by default in DOS INT10 video modes 0x00 - 0x0E that were introduced
prior to VGA to enable interoperability of programs that use CGA, EGA,
and early text modes to work with VGA video cards, such as the one
emulated by Spice86.

Prior to this commit, we already had the correct definition of the
WriteProtect bit in the VerticalSyncEndRegister, but it wasn't being
used anywhere. It was simply ignored. This worked correctly for games
that used VGA compatible video modes, and for some games that used
earlier EGA compatible video modes but didn't set any of the reserved
bits that were later used by VGA, but not for all earlier games.

To fix it, I made the CRT Controller properly respect the Vertical Sync
End register's Write Protect bit, which in turn enables compatibility
with more games that use EGA (and possibly CGA - although that remains
untested). I also added more detail to some of the comments and log
messages that I used to debug this problem, which will hopefully help
anyone looking at it in the future to pick it up more quickly.