Respect EGA/VGA read mode 1 contributions per plane in the video renderer
According to the EGA TRG's Theory of Operation, the Graphics Controller
Register's Read Mode 1 is intended to provide hardware assistance to
rapidly scan screen memory, find pixels that match a given color, and
provide the result to software in the Color Compare register. Our
implementation of this feature was *almost* correct, but we didn't pack
the bits correctly per-plane, so the color would *never* match the one
requested by software.

The fix was pretty simple: get only the bit for the current pixel in
each latch first, then shift it to the right position to construct the
full 4 bit value that makes up the pixel. This is repeated for all 8
pixels. (Previously we always got the first bit in the latch, for the
first pixel, and shifted it to the right position for the current pixel.
This meant that we repeated the color of the first pixel multiple
times.)

This fixes the rendering glitch in Cadaver reported in #1180. I also
regression tested it with Exodus, which uses the same EGA Video Mode
0x0D as Cadaver (but unlike Cadaver, doesn't use Read Mode 1), and there
aren't any regressions. I also regression tested it with Dune, which
similarly still renders correctly.